[WIP] KDL Compiler
This adds a WIP KDL compiler in Go, based losely on the learnings from
the Rust prototype.

Unlike the prototype, it functions in two phases -- first, it lexes into
tokens (including comment tokens -- more on that below), and then it
parses into a losely-typed AST.

Note that the losely-typed AST is *not* CKDL -- subsequent passes are
needed to resolve marker definitions, turn type modifier lists into
concrete types, etc.  Separating out those phases makes the parsing
logic a bit cleaner and easier to follow.

We use a split lexer-parser to potentially? make it easier to allow
less-destructive formatting (i.e. no loss of order, etc).  Furthermore,
this shows spots where we were previously allowing keywords to be used
as not-keywords.  Instead, we now support raw identifier syntax
(surrounded by backticks) as an escape hatch.

The parser itself uses a context-based system to keep track of source
spans, as well as provided context as to what it's doing.  This lets us
produce detailed error messages and locations (still need to do more
work to make the output a *little* nicer looking though).  Source spans
are carried over into the AST as well, so that we can generate span
information in the final CDKL serialization for downstream tooling.