Redo imports and type-checking
Old stuff is in the flows directory, not used any more, keeping around
for a bit longer to reference while restoring all the validation.

Basically:

Only nested name validation and modifier parsing is done on the AST now.
After that, the AST is converted to IR, joined with imported IR, and
typechecking & reference resolution is done on the IR instead.

This necessesitated introducing mechanics for tracking spans from IR,
but this isn't completely in place yet.  It uses a similar system to
proto, where we keep track of "paths" (proto tag numbers and repeated
indexes).  There's a helper system to make this easier in toir, a
similar system probably should exist in common to help keep track of
paths for reporting errors back.